add orders show funtion
