+++ app/assets/stylesheets/orders.scss
@@ -1,3 +1,3 @@
-// Place all the styles related to the order controller here.
+// Place all the styles related to the orders controller here.
 // They will automatically be included in application.css.
 // You can use Sass (SCSS) here: http://sass-lang.com/
