--- app/helpers/order_helper.rb
@@ -1,2 +1,0 @@
-module OrderHelper
-end
