+++ app/helpers/orders_helper.rb
@@ -1,0 +1,2 @@
+module OrdersHelper
+end
