+++ app/views/orders/show.html.erb
@@ -1,0 +1,64 @@
+<div class="row">
+  <div class="col-md-12">
+
+    <h2> 订单明细 </h2>
+
+    <table class="table table-bordered">
+      <thead>
+        <tr>
+          <th width="80%">商品明细</th>
+          <th>单价</th>
+        </tr>
+      </thead>
+      <tbody>
+
+        <% @product_lists.each do |product_list| %>
+          <tr>
+            <td>
+              <%= product_list.product_name %>
+            </td>
+            <td>
+              <%= product_list.product_price %>
+            </td>
+          </tr>
+        <% end %>
+
+      </tbody>
+    </table>
+
+    <div class="total clearfix">
+      <span class="pull-right">
+        总计 <%= @order.total %> CNY
+      </span>
+    </div>
+
+     <hr>
+
+     <h2> 寄送资讯 </h2>
+
+     <table class="table table-striped table-bordered">
+      <tbody>
+        <tr>
+          <td>
+            订购人
+          </td>
+        </tr>
+        <tr>
+          <td>
+            <%= @order.billing_name %> - <%= @order.billing_address %>
+          </td>
+        </tr>
+        <tr>
+          <td>
+            收件人
+          </td>
+        </tr>
+        <tr>
+          <td>
+            <%= @order.shipping_name %> - <%= @order.shipping_address %>
+          </td>
+        </tr>
+      </tbody>
+    </table>
+  </div>
+</div>
